hourly average for all devices in 2022
Hi Andrew, this is the hourly average for all devices in 2022. Do you think this format is appropriate/suitable for the following analysis? If so, I will continue to manage for 2023 without uploading the original files, which are too many, and then calculate daily, and monthly averages for them, after which we just need pretty plots to visualize them.